server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.0.0
torchvision>=0.15.0
numpy==1.24.3
pyarrow==14.0.1
datasets==2.14.0
modelscope==1.11.0
transformers>=4.35.0
Pillow>=10.0.0
faiss-cpu>=1.7.4
python-multipart>=0.0.6
addict>=2.4.0
simplejson>=3.19.0
oss2>=2.17.0
sortedcontainers>=2.4.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=4.25.0
bitsandbytes>=0.41.0
qwen-vl-utils
transformers_stream_generator
tiktoken
einops
scipy
